Outsource HG instantiation
